refactor: extract SuggestionAutocomplete component
Moves the inline SuggestionAutocomplete dropdown from App.jsx into
src/components/SuggestionAutocomplete.jsx. Theme variables (cardBg,
borderClass, textPrimary, hoverBg) are now passed as props to all
five call sites.
