+++ src/App.jsx
@@ -22,6 +22,7 @@ import { HabitRing, MiniHabitRing } from './components/HabitRing.jsx';
 import FrameEditor from './components/FrameEditor.jsx';
 import QuickAddFrameForm from './components/QuickAddFrameForm.jsx';
 import SmartSchedulePanel from './components/SmartSchedulePanel.jsx';
+import SuggestionAutocomplete from './components/SuggestionAutocomplete.jsx';
 
 // Encode a string that may contain non-ASCII characters as Base64.
 // btoa() throws InvalidCharacterError for codepoints > 255 (CJK, emoji, etc.).
@@ -10539,46 +10540,6 @@ const DayPlanner = () => {
     }
   }, [dataLoaded, hasZeroRealTasks]);
 
-  // Autocomplete dropdown component for tags, dates, and times
-  const SuggestionAutocomplete = ({ suggestions, selectedIndex, onSelect }) => {
-    if (suggestions.length === 0) return null;
-
-    const getIcon = (type) => {
-      switch (type) {
-        case 'date': return <Calendar size={14} className="flex-shrink-0" />;
-        case 'deadline': return <Calendar size={14} className="flex-shrink-0" />;
-        case 'time': return <Clock size={14} className="flex-shrink-0" />;
-        case 'duration': return <Clock size={14} className="flex-shrink-0" />;
-        case 'priority': return <AlertCircle size={14} className="flex-shrink-0" />;
-        default: return <Hash size={14} className="flex-shrink-0" />;
-      }
-    };
-
-    return (
-      <div className={`absolute top-full left-0 mt-1 ${cardBg} rounded-lg p-1 z-50 shadow-xl border ${borderClass} min-w-[160px] max-h-40 overflow-y-auto`}>
-        {suggestions.map((suggestion, index) => (
-          <button
-            key={`${suggestion.type}-${suggestion.value}-${index}`}
-            type="button"
-            onClick={(e) => {
-              e.preventDefault();
-              e.stopPropagation();
-              onSelect(suggestion);
-            }}
-            onMouseDown={(e) => e.preventDefault()} // Prevent blur before click
-            className={`w-full text-left px-2 py-1.5 rounded text-sm flex items-center gap-2 ${
-              index === selectedIndex
-                ? 'bg-blue-500 text-white'
-                : `${textPrimary} ${hoverBg}`
-            }`}
-          >
-            {getIcon(suggestion.type)}
-            <span className="truncate">{suggestion.display}</span>
-          </button>
-        ))}
-      </div>
-    );
-  };
 
   // Compute array of visible dates based on selectedDate and visibleDays
   const visibleDates = useMemo(() => {
@@ -19405,6 +19366,10 @@ const DayPlanner = () => {
                                         suggestions={suggestions}
                                         selectedIndex={selectedSuggestionIndex}
                                         onSelect={(suggestion) => applySuggestionForEdit(suggestion, editingInputRef.current, true)}
+                                        cardBg={cardBg}
+                                        borderClass={borderClass}
+                                        textPrimary={textPrimary}
+                                        hoverBg={hoverBg}
                                       />
                                     )}
                                   </div>
@@ -19899,6 +19864,10 @@ const DayPlanner = () => {
                                               suggestions={suggestions}
                                               selectedIndex={selectedSuggestionIndex}
                                               onSelect={(suggestion) => applySuggestionForEdit(suggestion, editingInputRef.current, false)}
+                                              cardBg={cardBg}
+                                              borderClass={borderClass}
+                                              textPrimary={textPrimary}
+                                              hoverBg={hoverBg}
                                             />
                                           )}
                                         </div>
@@ -20689,6 +20658,10 @@ const DayPlanner = () => {
                                                   suggestions={suggestions}
                                                   selectedIndex={selectedSuggestionIndex}
                                                   onSelect={(suggestion) => applySuggestionForEdit(suggestion, editingInputRef.current, false)}
+                                                  cardBg={cardBg}
+                                                  borderClass={borderClass}
+                                                  textPrimary={textPrimary}
+                                                  hoverBg={hoverBg}
                                                 />
                                               )}
                                             </div>
@@ -20759,6 +20732,10 @@ const DayPlanner = () => {
                                                   suggestions={suggestions}
                                                   selectedIndex={selectedSuggestionIndex}
                                                   onSelect={(suggestion) => applySuggestionForEdit(suggestion, editingInputRef.current, false)}
+                                                  cardBg={cardBg}
+                                                  borderClass={borderClass}
+                                                  textPrimary={textPrimary}
+                                                  hoverBg={hoverBg}
                                                 />
                                               )}
                                             </div>
@@ -23125,6 +23102,10 @@ const DayPlanner = () => {
                     suggestions={suggestions}
                     selectedIndex={selectedSuggestionIndex}
                     onSelect={applySuggestionForNewTask}
+                    cardBg={cardBg}
+                    borderClass={borderClass}
+                    textPrimary={textPrimary}
+                    hoverBg={hoverBg}
                   />
                 )}
                 {/* AI duration + tag suggestion pill */}
